Added function declaration and moved measurement entry struct to sensor.h
